crc32 should use unsigned int (now it works on amd64)

+++ zipscript/src/crc.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-unsigned long crc32_table[256] = {
+unsigned int crc32_table[256] = {
   0x00000000, 0x77073096, 0xee0e612c, 0x990951ba,
   0x076dc419, 0x706af48f, 0xe963a535, 0x9e6495a3,
   0x0edb8832, 0x79dcb8a4, 0xe0d5e91e, 0x97d2d988,
@@ -70,10 +70,10 @@ unsigned long crc32_table[256] = {
   0xb40bbe37, 0xc30c8ea1, 0x5a05df1b, 0x2d02ef8d,
  };
 
-unsigned long calc_crc32(char *f) {
+unsigned int calc_crc32(char *f) {
  FILE		*in;
  unsigned char	*buf;
- unsigned long	crc;
+ unsigned int	crc;
  size_t		i, j;
  int		k;
